requirements.txt
fastapi
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
alembic